### 2025-06-05 #8354 @editorialbot
[PRE REVIEW]: MatrixFuns.jl: Matrix functions in Julia
**Submitting author:** @xuequan818 (<a href="http://orcid.org/0000-0003-3349-9517">Xue Quan</a>)
**Repository:** https://github.com/xuequan818/MatrixFuns.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
