+++ paper/paper.bib
@@ -48,6 +48,7 @@
 	title     = {{Functions of Matrices: Theory and Computation}},
 	publisher = {SIAM},
 	address   = {Philadelphia},
+	doi 	  = {10.1137/1.9780898717778},
 	year      = {2008}
 }
 
