+++ paper/paper.md
@@ -40,7 +40,7 @@ The basic principle of the Schur-Parlett algorithm is as follows. First, one per
 
 Our algorithm attempts to find a partition of the eigenvalues of $A$ (computed using a Schur decomposition) into blocks that are well-separated. The diagonal blocks are then computed using Taylor series, and the Parlett recursion is used to fill out the off-diagonal blocks. The partitioning aims to find small blocks (so that low-order Taylor series can be used) that are well-separated (so that the Parlett recursion is well-conditioned)
 
-To find the partition, we start by partitioning the set of eigenvalues $\Lambda$ into disjoint clusters $\Lambda_i$ such that the distance between two such clusters is at least `sep`, where `sep` is a user-definable parameter. We then check if the partition is acceptable by estimating the error in all the clusters; if the estimated error is acceptable, we accept the partition; if not, we split the unacceptable clusters further by applying the partitioning algorithm recursively to each unacceptable $\Lambda_i$. We estimate the error in a cluster $\Lambda_i$ of diameter $d_i$ as $\textrm{err}_i = \left(\frac{d_i}{\textrm{scale}} \right)^{\textrm{max_deg}+1}$. We accept a cluster if $\textrm{err}_i \lt \varepsilon/\textrm{sep}$. This choice is made to balance the error originating from the Taylor expansion within a cluster $\textrm{err}_i$ with the error incurred by the use of the Parlett recursion $\varepsilon/\textrm{sep}$.
+To find the partition, we start by partitioning the set of eigenvalues $\Lambda$ into disjoint clusters $\Lambda_i$ such that the distance between two such clusters is at least `sep`, where `sep` is a user-definable parameter. We then check if the partition is acceptable by estimating the error in all the clusters; if the estimated error is acceptable, we accept the partition; if not, we split the unacceptable clusters further by applying the partitioning algorithm recursively to each unacceptable $\Lambda_i$. We estimate the error in a cluster $\Lambda_i$ of diameter $d_i$ as $\textrm{err}_i = \Big(\frac{d_i}{\textrm{scale}} \Big)^{\textrm{max_deg}+1}$. We accept a cluster if $\textrm{err}_i \lt \varepsilon/\textrm{sep}$. This choice is made to balance the error originating from the Taylor expansion within a cluster $\textrm{err}_i$ with the error incurred by the use of the Parlett recursion $\varepsilon/\textrm{sep}$.
 
 Therefore, our algorithm has the following parameters: 
 - `scale`, the characteristic scale of variations of $f$, set to $1$ by default.
