+++ paper/paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 The computation of matrix functions (i.e., $f(A)$ for $A$ a $n \times n$ matrix and $f : \mathbb{C} \to \mathbb{C}$) and their Fréchet derivatives plays a crucial role in many fields of science [@Higham2008], and in particular in electronic structure calculations within density functional theory and response calculations. For Hermitian $A$, computing $f(A)$ can be done efficiently and stably by diagonalization. In the non-normal case, however, diagonalization is unstable and alternative schemes have to be used. Even in the Hermitian case, the evaluation of Fréchet derivatives requires (high-order) divided differences, which by Opitz's formula [@deBoor2005] is equivalent to the exact computation of $f(A)$ for non-normal $A$.
 
-In this work, we develop `MatrixFuns.jl` a Julia package [@julia] to provide the robust computation of matrix functions for arbitrary square matrices and higher-order Fréchet derivatives for Hermitian matrices. This package is tailored towards high accuracy with relatively small matrices and relatively complicated functions $f$. Our work is based on the Schur-Parlett algorithm [@DaviesHigham03,HighamMohy10], with the following modifications:
+In this work, we develop `MatrixFuns.jl` a Julia package [@julia] to provide the robust computation of matrix functions for arbitrary square matrices and higher-order Fréchet derivatives for Hermitian matrices. This package is tailored towards high accuracy with relatively small matrices and relatively complicated functions $f$. Our work is based on the Schur-Parlett algorithm [@DaviesHigham03;@HighamMohy10], with the following modifications:
 - It supports functions that are discontinuous, or have sharp variations.
 - It does not require the computation of arbitrary-order derivatives of $f$.
 - It exploits existing special-purpose methods for computing matrix functions (e.g. for functions involving exponentials or logarithms) when they exist.
@@ -36,7 +36,7 @@ In this work, we develop `MatrixFuns.jl` a Julia package [@julia] to provide the
 
 # Methods
 ## Matrix functions
-The basic principle of the Schur-Parlett algorithm is as follows. First, one performs a Schur decomposition to reduce to the case of an upper triangular matrix. Then, one uses the Parlett recursion, which for a block matrix $A = \begin{pmatrix}A_{11}&A_{12}\\0&A_{22}\end{pmatrix}$ expresses $B = f(A)$ as $B_{11}=f(A_{11})$, $B_{22}=f(A_{22})$ and $B_{12}$ given by the solution of the Sylvester equation $A_{11} B_{12} - B_{12} A_{22} = B_{11}A_{12}-A_{12}B_{22}$. In principle, this can be used to compute $f(A)$ by a recursion, but the Sylvester equation becomes ill-conditioned when $A_{11}$ and $A_{22}$ do not have well-separated eigenvalues. In this case, one can use Taylor series, as proposed in [@DaviesHigham03,HighamMohy10], but this has the disadvantage of requiring arbitrarily many derivatives of $f$, which might be impractical in some applications (e.g. when the function is not analytic, or has sharp variations).
[…]
