### 2025-06-16 #8396 @shravanngoswamii
Just a suggestion (personal opinion): As of now, there is no registered package in the Julia Registry with the name *MatrixFunctions.jl*, so I think using this name would be much better from user's perspective. Alternatively, if shorter names are preferred, perhaps *MatrixFuncs.jl* would work, as "Funcs" represents "Functions" more clearly than "funs." I found [MatrixFunctions.jl](https://github.com/aravindh-krishnamoorthy/MatrixFunctions.jl), but it's not well developed, active, or registered.
