### 2025-06-16 #8396 @danielskatz
👋 @vlc1 - as a friendly reminder, you should use the command `@editorialbot generate my checklist` to create your review checklist. @editorialbot commands need to be the first thing in a new comment.
