### 2025-06-17 #8396 @xuequan818
> Just a suggestion (personal opinion): As of now, there is no registered package in the Julia Registry with the name _MatrixFunctions.jl_, so I think using this name would be much better from user's perspective. Alternatively, if shorter names are preferred, perhaps _MatrixFuncs.jl_ would work, as "Funcs" represents "Functions" more clearly than "funs." I found [MatrixFunctions.jl](https://github.com/aravindh-krishnamoorthy/MatrixFunctions.jl), but it's not well developed, active, or registered.

Thank you very much for your suggestion. I completely agree with your point. However, changing the name of an already registered package would require registering it as a new package, and I am not sure how this might affect the JOSS review and submission process. For reference, I also noticed that [StatsFuns.jl](https://github.com/JuliaStats/StatsFuns.jl) uses “Funs” to denote “Functions”.
