### 2025-06-17 #8396 @danielskatz
A name change wouldn't hurt the JOSS process.
