### 2025-07-02 #8396 @shravanngoswamii
@xuequan818 Can you please add a small note or something for contributors so we can check the Community guidelines in checklist
