+++ docs/src/background/limit.md
@@ -1,3 +1,3 @@
-# Limilations
+# Limitations
 
 The computation for the default separation parameter can be numerically unstable, which is an inherent limitation of the Schur-Parlett algorithm.
