### 2025-07-06 #8396 @shravanngoswamii
Hello @xuequan818,
MatrixFuns.jl is a very nice package -- great work! I just have a couple of suggestions:
- Consider adding some basic benchmarks (e.g., runtime vs. matrix size for functions like `exp`, `erf`).
- It might be helpful to clarify any scalability limits in the documentation.

I've also submitted a small PR: https://github.com/xuequan818/MatrixFuns.jl/pull/22

Overall, the paper is well-written and the package is a valuable contribution to the Julia ecosystem. I'm concluding my review here, but feel free to reach out if there's anything else I can help with!
