### 2025-07-21 f550375
fix taylor bug
