+++ .github/workflows/CI.yml
@@ -24,7 +24,8 @@ jobs:
       matrix:
         version:
           - '1.10'
-          - 'pre'
+          - '1'
+          - 'nightly'
         os:
           - ubuntu-latest
         arch:
