+++ .github/workflows/CI.yml
@@ -25,7 +25,6 @@ jobs:
         version:
           - '1.10'
           - '1'
-          - 'nightly'
         os:
           - ubuntu-latest
         arch:
