### 2025-07-27 #8396 @danielskatz
👋 @vlc1 - I've been on vacation for 2 weeks and am now trying to catch up. Can you let me know where this is from your point-of-view, and what the next step or blocker is for you?
