### 2025-08-13 #8396 @vlc1
Dear @xuequan818

Very nice work, thanks for this nice contribution to the Julia ecosystem. I highlighted below a couple of minor typos in the manuscript, @danielskatz it's all good on my end, thanks!

- The bulleted lists in the abstract of the paper (lines 56 to 64, where the inputs are documented) are not rendered correctly in the PDF.
- A typo line 73: Talor instead of Taylor.
- A couple of things to modify in the references:

	1. `Davies, P. I.` instead of `Davies, Philip I.`
	2. `de Boor, C.` instead of de `Boor, C. de.`
