### 2025-08-14 823dba0
add benchmark and limitations in doc
