### 2025-08-14 db41878
fix errors in paper
