+++ paper/paper.bib
@@ -11,7 +11,7 @@
 
 
 @article{DaviesHigham03,
-	author  = {Philip. I. Davies and Nicholas J. Higham},
+	author  = {Philip I. Davies and Nicholas J. Higham},
 	journal = {SIAM Journal on Matrix Analysis and Applications},
 	number  = {2},
 	pages   = {464-485},
@@ -33,7 +33,7 @@
 }
 
 @article{deBoor2005,
-	author    = {Carl de Boor},
+	author    = {de Boor, Carl},
 	journal   = {Surveys in Approximation Theory},
 	pages     = {46-69},
 	publisher = {Technion},
