+++ paper/paper.bib
@@ -33,7 +33,7 @@
 }
 
 @article{deBoor2005,
-	author    = {de Boor, Carl},
+	author    = {{de Boor}, Carl},
 	journal   = {Surveys in Approximation Theory},
 	pages     = {46-69},
 	publisher = {Technion},
