### 2025-08-17 #8396 @xuequan818
> 👋 [@xuequan818](https://github.com/xuequan818) - I've proofread the paper, and made suggestions in [xuequan818/MatrixFuns.jl#24](https://github.com/xuequan818/MatrixFuns.jl/pull/24). Please merge this, or let me know what you disagree with.
>
> I also note that there is no Acknowledgments section in the paper. Should there be, to acknowledge funders or others?

Thank you very much! I have merged this. There is nothing to acknowledge from us.
