### 2025-08-18 c34da91
remove unnecessary codes
