+++ v1.0.3/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"api/#Matrix-Functions-API","page":"API","title":"Matrix Functions API","text":"","category":"section"},{"location":"api/#Matrix-variable-functions-f(A::AbstractMatrix)::AbstractMatrix","page":"API","title":"Matrix-variable functions f(A::AbstractMatrix)::AbstractMatrix","text":"","category":"section"},{"location":"api/#MatrixFuns.mat_fun","page":"API","title":"MatrixFuns.mat_fun","text":"mat_fun(f, A; scale, sep, max_deg, color, ε, checknative)\n\nCompute the matrix-variable function f(A). Return a matrix.\n\nf a general scalar function, and called as f(x).\n\nA an arbitrary square matrix.\n\nscale the scaling of the Talyor series error, is used to control the spread of each splitting cluster. By default, scale = 1.\n\nsep the initial separation distance, to split the eigenvalues of A into clusters that satisfy (i) min{|λ - μ|: λ ∈ clp, μ ∈ clq, p ≠ q} > sep.\t\t (ii) for clp with |clp| > 1, ∀ λ ∈ clp, ∃ μ ∈ clp and μ ≠ λ, s.t. |λ - μ| ≤ sep.  Here cl_p is the cluster with index p. By default, sep = 0.1*scale.\n\nmax_deg the maximum Taylor series order for the diagonal blocks computation.\n\ncolor(x::Number)::Integer an index mapping function.  This is to ensure all eigenvalues within a cluster have the same color index.  By default, the color function maps x to 1.  For discontinuous functions, users can assign a distinct color to each continuous interval.  E.g., for the sign function, users can customize the color mapping using  color = x -> Int(real(sign(x)))\n\nε the target accuracy, set to machine accuracy by default.\n\nchecknative check f is a Julia native function.\n\n\n\n\n\n","category":"function"},{"location":"api/#Divided-Differences-of-f(x::Real)::Number","page":"API","title":"Divided Differences of f(x::Real)::Number","text":"","category":"section"},{"location":"api/#MatrixFuns.div_diff","page":"API","title":"MatrixFuns.div_diff","text":"div_diff(f, x::AbstractVector; kwargs...)\ndiv_diff(f, x::Tuple; kwargs...)\ndiv_diff(f, x...; kwargs...)\n\nReturn the divided difference f[x_0,x_1,...,x_n], assuming f is called as f(x).  See mat_fun for a description of possible keyword arguments.\n\n\n\n\n\n","category":"function"},{"location":"api/#Fréchet-derivatives-of-f(A::AbstractMatrix)::AbstractMatrix","page":"API","title":"Fréchet derivatives of f(A::AbstractMatrix)::AbstractMatrix","text":"","category":"section"},{"location":"api/#MatrixFuns.mat_fun_frechet","page":"API","title":"MatrixFuns.mat_fun_frechet","text":"mat_fun_frechet(f, H::AbstractMatrix, h::Vector{AbstractMatrix}; kwargs...)\nmat_fun_frechet(DD_F, Ψ::AbstractMatrix, h::Vector{AbstractMatrix})\n\nReturn the n-th order Fréchet derivative d^nf(H)h[1]…h[n], assuming f is called as f(x). See mat_fun for a description of possible keyword arguments.\n\n\n\n\n\n","category":"function"},{"location":"api/#Implementation-details","page":"API","title":"Implementation details","text":"","category":"section"},{"location":"api/#Matrix
[…]
