### 2025-08-19 #8396 @xuequan818
@danielskatz Here is the information for the archived version.

- version tag: v1.0.3
- Zenodo link: https://zenodo.org/records/16899270
- DOI: https://doi.org/10.5281/zenodo.16899270

Please let me know if anything is missing.
