### 2025-08-19 #8396 @shravanngoswamii
Congratulations @xuequan818! Your contribution is a valuable addition to the Julia community. Many thanks to @danielskatz — this was my first review for JOSS, and I really enjoyed it. The JOSS review pipeline is great — smooth, transparent, and well-organized!
